Add (broken) helper function for ICD/O -> NCIT
